+++ mllaunchpad/config.py
@@ -18,6 +18,30 @@ logger = logging.getLogger(__name__)
 
 CONFIG_DEFAULT = "./LAUNCHPAD_CFG.yml"
 CONFIG_ENV = os.environ.get("LAUNCHPAD_CFG", CONFIG_DEFAULT)
+required_config = {
+    # datasources and datasinks are optional
+    "model_store": {"location": {}},
+    "model": {
+        "name": {},
+        "version": {},
+        "module": {},
+        # custom model keys are optional
+    },
+    # api: is optional because using mllaunchpad's WSGI/Flask API is optional
+    # (e.g. when using mllaunchpad's convenience functions in creating Azure functions).
+    # You are free to add more custom settings in the config file. For an example, see
+    # https://github.com/schuderer/mllaunchpad-template
+}
+
+
+def validate_config(config_dict, required, path=""):
+    for item in required:
+        path_start = (path + ":") if path else ""
+        if item not in config_dict:
+            raise ValueError(
+                "Missing key in config file: {}".format(path_start + item)
+            )
+        validate_config(config_dict[item], required[item], path_start + item)
 
 
 def get_validated_config(filename=CONFIG_ENV):
@@ -45,17 +69,7 @@ def get_validated_config(filename=CONFIG_ENV):
     with open(filename) as f:
         y = yaml.load(f, Loader)
 
-    req_props = ["model", "api"]
-
-    for prop in req_props:
-        if prop not in y:
-            raise ValueError(
-                "Configuration file {} does not contain {} property.".format(
-                    filename, prop
-                )
-            )
-
-    # TODO: should completely validate config structure when stable
+    validate_config(y, required_config)
 
     logger.debug("Configuration loaded and validated: %s", y)